requirements.txt
pandas
numpy
pyarrow
scikit-learn
fastapi
uvicorn
python-dotenv
kagglehub
ipykernel
fastparquet
apscheduler
django